requirements.txt
colorama==0.4.6
Faker==37.5.3
fastapi==0.116.1
langchain==0.3.27
langgraph==0.6.3
passlib==1.7.4
protobuf==6.31.1
pydantic==2.11.7
pydantic_settings==2.10.1
python_jose==3.5.0
Requests==2.32.4
SQLAlchemy==2.0.42
uvicorn==0.35.0
psycopg2-binary==2.9.10
google-auth==2.40.3
pydantic[email]==2.11.7
langchain-openai==0.3.28